Remove more refernces to long
The sftp code had been using whether a value was a long or int
to signal whether it should be encoded as a 64- or 32-bit int
on the wire.

We still need a way of doing this apparently, so introduce a trivial
subclass of int to signal that we want to be 64-bit.